Refactor TransactionActivityTest to use TestInstallIn Module
